+++ src/smagic.c
@@ -4243,7 +4243,7 @@ void spell_spirit_jump(int level, P_char ch, char *arg, int type, P_char victim,
   }
 
 //  distance = (level * 27) / 5 + (IS_SPIRITUALIST(ch) ? 15 : 0);
-  distance = (int)(level * (int)get_property("spell.dim.perlevel.modifier", 1.35));
+  distance = (int)(level * get_property("spell.dim.perlevel.modifier", 1.35));
   if(GET_SPEC(ch, CLASS_SHAMAN, SPEC_SPIRITUALIST))
     distance += 15;
